Program.cs
Current version of the main code. Please note that I got ahead of myself in finishing some work on this project, so my commenting will be a little behind. I will amend this with the next update, after I review proper commenting and naming styles that is used for C# (I'm still accustomed to what I learned for C++ in my class, so I'm unsure if they're different or not).

To-Do:
   - Clean-up and comment code (Some code may poorly organized, or there may be better methods usable)
   - Finish Add/Remove function (missing ability to remove a character, 'Add character' case needs code to be cleaned up)
   - (Later on) Ability to reset text file data.