Program.cs
Updated 7/30/17

The application, upon execution, now sets the console window to a default size, one that should make it easier for users to read the material, as opposed to having to set it themselves. 

Additionally, if there is no directory found for the CharacterData files, it creates a CharacterData directory in the public documents folder. 

The last of this version of the applications features have been implemented, so, this program is now feature complete. For the sake of practicing other techniques, some things still need to be re-written or commented. See below.

Much code has been re-written, mainly to avoid having everything in a try-catch block.

Still Need To:
   - Change loading path (Currently at CharacterData.txt, for testing purposes)
   - Replace certain try-catch blocks (Apparently it's bad practice to use this when Int32.TryParse() exists, for checking numerical input)
   - Properly comment and document code.

The above tasks should be finished over the next couple of days.