Program.cs
With this, I consider this project to be finished, unless I find issue with it at some point in the future. Code has been cleaned, documented, and commented. If I find an issue with it during use of the application, I will update the code, as well as the files on here. Anyone who has been following this, please feel free to point out any issues, or offer any advice on what I wrote.

This project took me about a month of working on and off of it, totaling about 1200 lines of code, including documentation and comments.

Thanks!